requirements.txt
google-auth-oauthlib==0.4.6
google-auth==2.3.3
requests==2.26.0
icalendar==4.0.7

